More fixed syscall return types and fixed sched affinity faking

+++ src/nautilus/syscalls/exit.c
@@ -3,7 +3,7 @@
 #include <nautilus/shell.h>
 
 
-int sys_exit(int exit_status)
+uint64_t sys_exit(uint64_t exit_status)
 {
   nk_thread_exit((void*)(long)exit_status);
 	return 0;
