+++ src/nautilus/syscalls/mprotect.c
@@ -4,7 +4,7 @@
 #define DEBUG(fmt, args...) DEBUG_PRINT("sys_mprotect: " fmt, ##args)
 #define INFO(fmt, args...) INFO_PRINT("sys_mprotect: " fmt, ##args)
 
-int sys_mprotect(int addr, int length, int prot) {
+uint64_t sys_mprotect(int addr, int length, int prot) {
   // "Works" by mmap setting liberal permissions
   DEBUG("This syscall does nothing\n");
   return 0;
